+++ tests/test_custom_checkers/test_infinite_loop_checker.py
@@ -12,45 +12,40 @@ class TestInfiniteLoopChecker(pylint.testutils.CheckerTestCase):
         src = """
             i, k, l = 0, 20, 40
             j = 10
-            while i < 100 and k < 21 or l < 40:
+            while i < 100 and k < 21 or l < 40: #@
                 j += 1
                 j = j - 1
         """
-        mod = astroid.parse(src)
 
-        node, *_ = mod.nodes_of_class(astroid.nodes.While)
+        node = astroid.extract_node(src)
 
         with self.assertAddsMessages(
             pylint.testutils.MessageTest(
-                msg_id="loop-condition-variable-unused",
+                msg_id="infinite-loop",
                 node=node.test,
             ),
             ignore_position=True,
         ):
             self.checker.visit_while(node)
-            self.checker.leave_while(node)
 
     def test_attr_not_updated(self) -> None:
         """Test that the checker correctly flags a while loop when no condition variable attributes are used in
         the loop body."""
         src = """
-        while 0 < self.attribute < 100:
+        while 0 < self.attribute < 100: #@
             attribute += 1
         """
 
-        mod = astroid.parse(src)
-
-        node, *_ = mod.nodes_of_class(astroid.nodes.While)
+        node = astroid.extract_node(src)
 
         with self.assertAddsMessages(
             pylint.testutils.MessageTest(
-                msg_id="loop-condition-variable-unused",
+                msg_id="infinite-loop",
                 node=node.test,
             ),
             ignore_position=True,
         ):
             self.checker.visit_while(node)
-            self.checker.leave_while(node)
 
     def test_nested_while_unused_var(self) -> None:
         """Test that the checker flags a nested while loop where none of its condition variables are used in its body"""
@@ -58,120 +53,121 @@ class TestInfiniteLoopChecker(pylint.testutils.CheckerTestCase):
         i = 0
         j = 10
         k = "David is cool!"
-        while i < 10:
-            while i < j:
+        while i < 10: #@
+            while i < j: #@
                 k += '!'
             i += 1
         """
 
-        mod = astroid.parse(src)
-
-        nodes = list(mod.nodes_of_class(astroid.nodes.While))
+        node, detected_node = astroid.extract_node(src)
 
         with self.assertAddsMessages(
             pylint.testutils.MessageTest(
-                msg_id="loop-condition-variable-unused",
-                node=nodes[1].test,
+                msg_id="infinite-loop",
+                node=detected_node.test,
             ),
             ignore_position=True,
         ):
-            for node in mod.nodes_of_class(astroid.nodes.While):
+            for node in node.nodes_of_class(astroid.While):
                 self.checker.visit_while(node)
-                self.checker.leave_while(node)
 
     def test_while_inside_func_unused_var(self) -> None:
         """Test that the checker flags infinite while loops inside function."""
         src = """
         def foo():
             i = 10
-            while i < 21:
+            while i < 21: #@
                 j = 19
         """
-        mod = astroid.parse(src)
-
-        node = list(mod.nodes_of_class(astroid.nodes.While))
+        node = astroid.extract_node(src)
 
         with self.assertAddsMessages(
             pylint.testutils.MessageTest(
-                msg_id="loop-condition-variable-unused",
-                node=node[0].test,
-            ),
-            ignore_position=True,
-        ):
-            for node in mod.nodes_of_class(astroid.nodes.While):
-                self.checker.visit_while(node)
-                self.checker.leave_while(node)
-
-    def test_subscript_notation(self) -> None:
-        """Test that the checker flags a while loop that uses a subscript annotated variable."""
-        src = """
-        lst = [1, 2, 3]
-        while self.lst[1] > 1:
-            lst[0] += 1
-        while lst[0] > 1:
-            self.lst[1] += 1
-        """
-
-        mod = astroid.parse(src)
-
-        node, *_ = mod.nodes_of_class(astroid.nodes.While)
-
-        with self.assertAddsMessages(
-            pylint.testutils.MessageTest(
-                msg_id="loop-condition-variable-unused",
+                msg_id="infinite-loop",
                 node=node.test,
             ),
             ignore_position=True,
         ):
             self.checker.visit_while(node)
-            self.checker.leave_while(node)
+
+    def test_subscript_notation(self) -> None:
+        """Test that the checker flags a while loop that uses a subscript annotated variable."""
+        src = """
+        lst = [1, 2, 3]
+        i = 0
+        while lst[1] > 1: #@
+            i += 1
+        """
+
+        node = astroid.extract_node(src)
+
+        with self.assertAddsMessages(
+            pylint.testutils.MessageTest(
+                msg_id="infinite-loop",
+                node=node.test,
+            ),
+            ignore_position=True,
+        ):
+            self.checker.visit_while(node)
+
+    def test_condition_variable_used_only_as_argument(self):
+        """Test that the checker does not flag while loops when the condition variable is passed to a function
+        or method."""
+        src = """
+         i = 0
+         while i < 20: #@
+            self.update(i)
+
+         while i < 20: #@
+            increment(i)
+         """
+
+        node1, node2 = astroid.extract_node(src)
+
+        with self.assertNoMessages():
+            self.checker.visit_while(node1)
+            self.checker.visit_while(node2)
 
     def test_multiple_while_pass(self) -> None:
         """Test that the checker does not flag non-infinite while loops."""
         src = """
         i = 0
         j = 10
-        while i < 10:
+        while i < 10: #@
             i += 2
-        while j < 20:
+        while j < 20: #@
             j += 1
         """
 
-        mod = astroid.parse(src)
+        node1, node2 = astroid.extract_node(src)
 
         with self.assertNoMessages():
-            for node in mod.nodes_of_class(astroid.nodes.While):
-                self.checker.visit_while(node)
-                self.checker.leave_while(node)
+            self.checker.visit_while(node1)
+            self.checker.visit_while(node2)
 
     def test_while_inside_func_pass(self) -> None:
         """Test that the checker does not flag non-infinite while loops inside function."""
         src = """
         def foo():
             i = 10
-            while i < 21:
+            while i < 21: #@
                 i += 1
                 j = 19
         """
-        mod = astroid.parse(src)
+        node = astroid.extract_node(src)
 
         with self.assertNoMessages():
-            for node in mod.nodes_of_class(astroid.nodes.While):
-                self.checker.visit_while(node)
-                self.checker.leave_while(node)
+            self.checker.visit_while(node)
 
     def test_while_true_enabled(self) -> None:
         """Test that loops with no condition variables is not flagged."""
         src = """
         x = 0
-        while True:
+        while True: #@
             x += 1
         """
 
-        mod = astroid.parse(src)
-
-        node, *_ = mod.nodes_of_class(astroid.nodes.While)
+        node = astroid.extract_node(src)
 
         with self.assertNoMessages():
             self.checker.visit_while(node)
-            self.checker.leave_while(node)